Restructuring classes
