+++ CodeChallenges/FifoAnimalShelter/Animal.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace FifoAnimalShelter
+{
+    public abstract class Animal
+    {    
+        
+            public abstract string Species { get; set; }
+            public abstract Animal Next { get; set; }
+            public abstract int UniqueID { get; set; }      
+            public abstract 
+
+        
+    }
+}
